Adding client-services/api-service and models file
