+++ src/models.ts
@@ -1,0 +1,4 @@
+export interface Kit {
+  labelId: string;
+  shippingTrackingCode: string;
+}
